feat: Added default help command

+++ data/plugins/clif/commands/$plugin delete.py
@@ -19,13 +19,17 @@ def export() -> dict:
 # Must be in every single command files.
 def execute(req_args : list[str], opt_args : dict = {}, suplementary : dict = {}) -> None:    
     plugin_name = req_args[0]
+    essential_plugin_list = ["clif"]
 
     formated_name : str = plugin_name.lower().replace(" ", "_")
     plugin_directory_path : str = global_settings["plugins_directory"] + (formated_name)
 
     if not os.path.exists(plugin_directory_path):
         terminal.print_err("CLIF.COMMAND.PLUGIN$DELETE.PLUGIN_PATH_NOT_FOUND", placeholders={"%pack_path%": plugin_directory_path})
-        return        
+        return
+    
+    if plugin_name in essential_plugin_list:
+        return
     
     try:
         shutil.rmtree(plugin_directory_path)
